refactor: move remaining mutations into hooks
