+++ apps/dashboard/src/app/(dashboard)/event/[id]/edit-card.tsx
@@ -1,14 +1,12 @@
 import { FC } from "react"
-import { trpc } from "../../../../utils/trpc"
 import { useEventWriteForm } from "../write-form"
 import { EventWriteSchema } from "@dotkomonline/types"
 import { useEventDetailsContext } from "./provider"
-import { useQueryNotification } from "../../../notifications"
-import { useEditEvent } from "../../../../modules/event/use-edit-event"
+import { useEditEventMutation } from "../../../../modules/event/use-edit-event-mutation"
 
 export const EventEditCard: FC = () => {
   const { event } = useEventDetailsContext()
-  const edit = useEditEvent()
+  const edit = useEditEventMutation()
   const FormComponent = useEventWriteForm({
     label: "Oppdater arrangement",
     onSubmit: (data) => {
